remove the "isa" binary operator
This is really frustrating!  The "isa" operator was introduced in v5.32
as an experiment.  In v5.36, it was made stable... but it turns out to
have a horrible bug, where using "isa" can kill the runtime's ability
to use the ->isa method.  This was found and fixed in v5.34...
partially!

I don't know if this feature is dead to me forever, but it's dead to me
for v5.36.