Linear: do not produce 0 character text nodes
We were sometimes producing Block Kit content with:

    { "type": "text", "text": "" }

…which is not legal.  Empty strings are forbidden.

This happened when we wanted to highlight "foo" in "foo bar baz",
specifically where "foo" appears *first* in the string.

We split on "foo" and then ended up highlighting the zero-width
substring in front of "foo", which got that block we don't like.

Let's just grep out all empty hunks.