InABox: box vpn omit tag in hostname by default
Currently when generating vpn config we are always including the tag in
the fqdn generated. eg 'user-bullsye.boxdomain.com'

This is fine until we bump the default version from bullseye -> bookworm
and people get understandably confused when their vpn doesn't work any
more.

On 'box create' without a /version or /tag passed in we setup dns for both
the 'user-tag.boxdomain.com' and 'user.boxdomain.com' (the latter was
the only thing we had before we added multiple box support).

So lets rely on the dns entry omitting the tag by default when invoking
'box vpn' without a /version or /tag switch