Reactor/Linear: Implement 'triage /all'
The 'triage' command by default shows only unassigned issues, in order
to assist with triaging tickets created or moved from other teams.

However, tickets we (in PLAT) create for triage through `synergy ++`
will have the creator assigned. This is fine but we want to be able to
query synergy for these tickets, so we do that with `synergy triage plat
/all`.

Ticket: PLAT-2422