Page: do not lose vertical whitespace
In 7307ef81 and its parent, Page changed to split the message on \s+ and
then join later with \N{SPACE}.  This turned vertical whitespace into
horizontal whitespace.  This broke a test, but we didn't notice. 🙁

I'd rather preserve space, but that's a bit more work and I'm not very
motivated.  So, this just fixes the test and keeps blank lines working.