tests: reenable network-listening tests under GitHub Actions
These tests can sit listening forever unless terminated.  The problem (I
believe) is that yath was meant to catch a term signal, but it's running
as pid 1, which has special rules.  We need to put an init in the way,
to we have added "--init" to the container options in the underlying
dzil-action action.