Reactor/InABox: Setup is now mandatory
Because of inabox changes, `fmdev mysetup` now needs to be run on
every deployed inabox.

So we make the ssh_key_id private key file mandatory and rejigger the
setup logic so that it always runs and passes the new "--no-custom"
argument to fmdev mysetup if the custom setup shouldn't run.