Synergy::BoxManager: inabox management outside the hub
Synergy is great and all, but sometimes it's nice to manage boxes from
outside of chatops.  Making that possible was the real reason behind
ever writing Dobby, and this commit realizes it.  This abstracts box
management (create, destroy, setup, power events) from the InABox
reactor.

The BoxManager needs to be given callbacks for chatops things like
sending replies.  In a console environment, these can be replaced with
something like "print".

There are a few reductions in functionality here, but I think they're
worth the benefit of abstraction.  I believe they are:

1.  The "use /force" message now says "use force" and leaves it up to
    the operator to know how to communicate that to the box manager.
2.  The "snapshot not here but is in region X" message drops the "but is
    in region X" part.  This can probably be brought back, but this
    situation should be very unlikely, and we're still much better off
    than the previous condition:  "Something weird happened."
3.  The (weirdly) slow power on/off/down operations no longer use a
    Slack reaction to indicate work, they just reply.
4.  When posting a snippet of problems running the setup program, only
    GitLab will be used.  The Slack code has some odd ordering issues,
    and was not being used in practice, anyway.