dist.ini: downgrade Dobby::BoxManager prereq
AutoPrereqs started picking up on Dobby::BoxManager a little while ago,
which meant this showed up as a prereq for the dist, which broke the CI
build, which does a "dzil build" and installs that -- the installers
started seeing Dobby::BoxManager, which was not available.

One fix would have been to leave it and move the before-ci step to
before-make or something.  But we already "downgrade" the Dobby::Client
prereq by removing it from the auto-created "runtime requires" into a
manual "runtime suggests".  So, this just does the same.  Everything
works.