Linear: retain the add_attachment_to_issue
This is just used to say "the Issue came from this Slack message", and
doesn't have to succeed.  Rather than wait for it with `await` and risk
breaking the rest of the process, we just ->retain the future so it will
(maybe) complete in the background.

(We could use await and try, but it would slow down response and it
doesn't seem worth it.)