Reactor/PagerDuty: give oncall to escalation
This creates the new special "escalation" target value for give-oncall command,
such as:

give oncall to escalation for 2 hours

which will figure out who to give oncall to automatically.

It will simply figure out who is currently on call in the second oncall
escalation level and give oncall to the first (usually only) person who
is on call for that.