Reactor/PagerDuty: give oncall: Try to figure out duration if not given
When no duration is given for `give oncall`, determine the shift end of
the current oncaller (that the oncall is being taken from) and place the
override until the end of the shift.

In the unlikely case that the shift has no end (is perpetual), it will
error out as before when no duration is given.