Reactor/PagerDuty: Resolve mine when handing over to escalation
`give oncall to escalation` is usually used as an emergency button if
the oncaller is unable to handle alerts due to personal reasons. They
may have acknowledged alerts at this time, which the escalation oncall
will then have to take over. In order to avoid letting the 30 minute
unacknowledge timer run out, we resolve these incidents and let them
retrigger. This should happen within 2 to 5 minutes based on our regular
monitor frequencies.

We don't need to do this for snoozed incidents, as we expect that those
won't require immediate attention. Definitively determining whether an incident
is snoozed is annoying (requires additional API call to get the incident
action log) so we just check for alerts due to unacknowledge within the
next 30 minutes which is morally equivalent.