Slack: simplify how the "slack" alt works
Previously, the "slack" alt to send_message could be a simple string or
a reference to a hash of arguments to chat.postMessage.  Now it is
either a simple string or a Slack::BlockKit::BlockCollection object.

The only argument we ever used was "blocks", which was meant to be the
kind of data you'd get by calling $block_collection->as_struct.  By
eliminating options other than "just blocks", and by requiring that
blocks are only in BlockCollection form, we make it easier to inspect or
combine blocks.