Report: combine Slack blocks from multiple reports
We used to just join the Slack alts from multiple reports with "\n",
which was fine when they were all text, but what if we want a report
producer wants to use BlockKit?  This would've just stuck HASH(0x...)
into the report, which is no good.

Instead, we now extract blocks from BlockCollections, then upgrade text
into blocks, then we put them all into a new BlockCollection and report
that.