GitLab: adapt (not perfectly) to request-do-not-assign workflow
We're changing our use of GitLab from "assign to request review" to "use
the review request system".  I think this is a change for the better,
and leans into letting GitLab's own interface work better.  (Like: we
don't want to store our important email in the trash folder.)

The frustrating down side is that there appears to be no practical API
access for "what is this reviewer's current review of this MR".  The UI
can display an icon for "has requested changes", but there is no
reliable way to answer this from the API.  (You can do things like look
for system-made comments that read "Requested changes", but this is not
efficient and feels likely to break in the future.)

So, the changes will now show code as "in review", without saying
"waiting on reviewer", if the MR is assigned to a reviewer and not
approved.  So, the question of "is the author or the reviewer blocking"
is not entirely clear from this report.

The good news is that /dashboard/merge_requests, in GitLab, seems to
collect all of this information into its interface.

When the API improves, I will improve this code further.