Linear: add a "no-self-triage teams" list
Teams in this list won't use triage as the default state for tasks.
When a member of one of these teams adds a task to that team, it will
use the team's default status (unless a status is supplied), not triage.

Also, if no specific team member was chosen, it will be assigned to the
creating user.